+++ unlock-js/src/__tests__/integration/walletServiceIntegration.test.js
@@ -1,7 +1,7 @@
 import WalletService from '../../walletService'
 import Web3Service from '../../web3Service'
 import locks from '../helpers/fixtures/locks'
-import deployErc1820 from '../helpers/erc1820'
+import { waitForContractDeployed } from '../helpers/waitForContractDeployed'
 
 let host,
   port = 8545
@@ -23,18 +23,12 @@ let provider = `http://${host}:${port}`
 // - withdrawFromLock
 
 // Increasing timeouts
-jest.setTimeout(15000)
+jest.setTimeout(300000)
 
 let accounts
 
 // Tests
 describe('Wallet Service Integration', () => {
-  beforeAll(async () => {
-    // TODO: consider moving this to the docker setup?
-    // Deploying erc1820 which is required for later versions of public Lock
-    await deployErc1820(provider)
-  })
-
   const versions = ['v0', 'v01', 'v02', 'v10', 'v11', 'v12']
   describe.each(versions)('%s', versionName => {
     let walletService, web3Service
@@ -104,9 +98,22 @@ describe('Wallet Service Integration', () => {
       describe.each(
         locks[versionName].map((lock, index) => [index, lock.name, lock])
       )('lock %i: %s', (lockIndex, lockName, lockParams) => {
-        let lock, lockAddress, lockCreationHash
+        let lock, expectedLockAddress, lockAddress, lockCreationHash
 
         beforeAll(async () => {
+          if (lockParams.currencyContractAddress) {
+            // Let's wait for erc20Address to be deployed
+            await waitForContractDeployed(
+              web3Service.provider,
+              lockParams.currencyContractAddress
+            )
+          }
+
+          expectedLockAddress = await web3Service.generateLockAddress(
+            accounts[0],
+            lockParams
+          )
+
           lockAddress = await walletService.createLock(
             lockParams,
             (error, hash) => {
@@ -119,6 +126,11 @@ describe('Wallet Service Integration', () => {
         it('should have yielded a transaction hash', () => {
           expect.assertions(1)
           expect(lockCreationHash).toMatch(/^0x[0-9a-fA-F]{64}$/)
+        })
+
+        it('should have deployed a lock at the expected address', async () => {
+          expect.assertions(1)
+          expect(lockAddress).toEqual(expectedLockAddress)
         })
 
         it('should have deployed the right lock version', async () => {